@@ -16,24 +16,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^11.0.5",
-    "@angular/cdk": "^11.0.2",
-    "@angular/common": "~11.0.2",
-    "@angular/core": "~11.0.2",
-    "@angular/forms": "~11.0.2",
-    "@angular/platform-browser": "~11.0.2",
-    "@angular/platform-browser-dynamic": "~11.0.2",
-    "@angular/router": "~11.0.2",
-    "@capacitor-community/admob": "^1.2.5",
-    "@capacitor-community/firebase-analytics": "^0.2.0",
-    "@capacitor-community/firebase-crashlytics": "^0.3.0",
-    "@capacitor/core": "2.4.4",
-    "@capacitor/ios": "^2.4.4",
+    "@angular/animations": "^12.0.2",
+    "@angular/cdk": "^12.0.2",
+    "@angular/common": "~12.0.2",
+    "@angular/core": "~12.0.2",
+    "@angular/forms": "~12.0.2",
+    "@angular/platform-browser": "~12.0.2",
+    "@angular/platform-browser-dynamic": "~12.0.2",
+    "@angular/router": "~12.0.2",
+    "@capacitor-community/firebase-analytics": "^1.0.0",
+    "@capacitor-community/firebase-crashlytics": "^1.0.0",
+    "@capacitor/app": "^1.0.0",
+    "@capacitor/browser": "^1.0.0",
+    "@capacitor/core": "^3.0.0",
+    "@capacitor/device": "^1.0.0",
+    "@capacitor/haptics": "^1.0.0",
+    "@capacitor/ios": "3.0.0",
+    "@capacitor/keyboard": "^1.0.0",
+    "@capacitor/status-bar": "^1.0.0",
+    "@capacitor/storage": "^1.0.0",
     "@ionic-native/core": "^5.0.0",
     "@ionic-native/splash-screen": "^5.0.0",
     "@ionic-native/status-bar": "^5.0.0",
     "@ionic-super-tabs/angular": "^7.0.8",
     "@ionic/angular": "^5.0.0",
+    "@logtail/node": "^0.1.7",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
     "@ngxs-labs/async-storage-plugin": "^0.1.1",
@@ -51,15 +58,15 @@
     "smoothscroll-polyfill": "^0.4.4",
     "sw-toolbox": "^3.6.0",
     "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.2",
-    "@angular/cli": "~11.0.2",
-    "@angular/compiler": "~11.0.2",
-    "@angular/compiler-cli": "~11.0.2",
-    "@angular/language-service": "~11.0.2",
-    "@capacitor/cli": "2.4.4",
+    "@angular-devkit/build-angular": "~12.0.2",
+    "@angular/cli": "~12.0.2",
+    "@angular/compiler": "~12.0.2",
+    "@angular/compiler-cli": "~12.0.2",
+    "@angular/language-service": "~12.0.2",
+    "@capacitor/cli": "^3.0.0",
     "@ionic/angular-toolkit": "^2.3.0",
     "@ngxs/devtools-plugin": "^3.7.1",
     "@types/chart.js": "^2.9.29",
@@ -70,7 +77,7 @@
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.1.1",
+    "karma": "~6.3.2",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~3.3.0",
@@ -78,7 +85,7 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.0.5"
+    "typescript": "~4.2.4"
   },
   "description": "An Ionic project"
 }
