@@ -1,4 +1,16 @@
 export const environment = {
-  production: true,
-  tabtUrl: 'http://juc.floca.be:3004'
-};
+  production: false,
+  tabtUrl: 'https://tabt-rest.herokuapp.com',
+  internalPages: {
+    register: 'redirect-register.html'
+  },
+  firebaseConfig: {
+    apiKey: 'putakeyhere',
+    authDomain: 'beping-196714.firebaseapp.com',
+    databaseURL: 'https://beping-196714.firebaseio.com',
+    projectId: 'beping-196714',
+    storageBucket: 'beping-196714.appspot.com',
+    messagingSenderId: '169133988850',
+    appId: '1:169133988850:web:f83e033f93de42ceb6696b',
+    measurementId: 'G-NLCDC9F2GF'
+  }};
