@@ -4,7 +4,7 @@
 
 export const environment = {
     production: false,
-    tabtUrl: 'http://juc.floca.be:3004',
+    tabtUrl: 'http://localhost:3004',
     internalPages: {
         register: 'redirect-register.html'
     },
@@ -27,4 +27,4 @@ export const environment = {
  * This import should be commented out in production mode because it will have a negative impact
  * on performance if an error is thrown.
  */
-// import 'zone.js/dist/zone-error';  // Included with Angular CLI.
+// import 'zone.js/plugins/zone-error';  // Included with Angular CLI.
